Move out of /lib because of deps, remove basic jwt validation
