+++ ansible_base/workload_identity/exceptions.py
